iterate through array to find specific element
I found out after completing the exercise that a boolean would be smart
to use instead of my "cheating" by telling index to not go past 8. Had I
used a boolean like "found" I could've run through the entire array and
if found is false then it would print the "not found" text.